refactor: RunOwner 枚举移至 Constants/Enums
